scripts/requirements.txt
IPython
pwntools
matplotlib
ipdb
pypcapkit
dpkt
